fix: show windows title bar on sign in screen
