+++ frontend/appflowy_flutter/lib/user/presentation/screens/sign_in_screen/desktop_sign_in_screen.dart
@@ -1,18 +1,18 @@
 import 'package:appflowy/core/frameless_window.dart';
 import 'package:appflowy/env/cloud_env.dart';
 import 'package:appflowy/generated/locale_keys.g.dart';
+import 'package:appflowy/shared/window_title_bar.dart';
 import 'package:appflowy/user/application/sign_in_bloc.dart';
 import 'package:appflowy/user/presentation/screens/sign_in_screen/widgets/widgets.dart';
 import 'package:appflowy/user/presentation/widgets/widgets.dart';
+import 'package:appflowy_editor/appflowy_editor.dart';
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flowy_infra_ui/flowy_infra_ui.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
 class DesktopSignInScreen extends StatelessWidget {
-  const DesktopSignInScreen({
-    super.key,
-  });
+  const DesktopSignInScreen({super.key});
 
   @override
   Widget build(BuildContext context) {
@@ -20,9 +20,12 @@ class DesktopSignInScreen extends StatelessWidget {
     return BlocBuilder<SignInBloc, SignInState>(
       builder: (context, state) {
         return Scaffold(
-          appBar: const PreferredSize(
-            preferredSize: Size(double.infinity, 60),
-            child: MoveWindowDetector(),
+          appBar: PreferredSize(
+            preferredSize:
+                Size.fromHeight(PlatformExtension.isWindows ? 40 : 60),
+            child: PlatformExtension.isWindows
+                ? const WindowTitleBar()
+                : const MoveWindowDetector(),
           ),
           body: Center(
             child: AuthFormContainer(
@@ -31,7 +34,7 @@ class DesktopSignInScreen extends StatelessWidget {
                   title: LocaleKeys.welcomeText.tr(),
                   logoSize: const Size(60, 60),
                 ),
-                const VSpace(30),
+                const VSpace(20),
 
                 // const SignInAnonymousButton(),
                 const SignInWithMagicLinkButtons(),
@@ -55,9 +58,9 @@ class DesktopSignInScreen extends StatelessWidget {
                     final type = state.loginType == LoginType.signIn
                         ? LoginType.signUp
                         : LoginType.signIn;
-                    context.read<SignInBloc>().add(
-                          SignInEvent.switchLoginType(type),
-                        );
+                    context
+                        .read<SignInBloc>()
+                        .add(SignInEvent.switchLoginType(type));
                   },
                 ),
 
@@ -85,20 +88,12 @@ class _OrDivider extends StatelessWidget {
   Widget build(BuildContext context) {
     return Row(
       children: [
-        const Flexible(
-          child: Divider(
-            thickness: 1,
-          ),
-        ),
+        const Flexible(child: Divider(thickness: 1)),
         Padding(
           padding: const EdgeInsets.symmetric(horizontal: 10),
           child: FlowyText.regular(LocaleKeys.signIn_or.tr()),
         ),
-        const Flexible(
-          child: Divider(
-            thickness: 1,
-          ),
-        ),
+        const Flexible(child: Divider(thickness: 1)),
       ],
     );
   }
